Linux targets for sockets.

+++ src/nw/SysSocket.h
@@ -50,7 +50,7 @@ namespace NetworkBase
 uint64_t htonll(uint64_t hostllong);
 uint32_t htonl(uint32_t hostlong);
 uint16_t htons(uint16_t hostshort);
-uint64_t ntohll(uint32_t netllong);
+uint64_t ntohll(uint64_t netllong);
 uint32_t ntohl(uint32_t netlong);
 uint16_t ntohs(uint16_t netshort);
